[ES|QL] Move more types to esql-types package (#219661)
## Summary

Moves the types from the esql plugin to the esql-types package

---------

(cherry picked from commit 99bd14cce8ab71794506f94b578f0f4eae65339d)
