+++ src/platform/packages/shared/kbn-esql-types/index.ts
@@ -9,6 +9,10 @@
 
 export {
   ESQLVariableType,
+  EsqlControlType,
+  VariableNamePrefix,
+  type ControlWidthOptions,
+  type ESQLControlState,
   type ESQLControlVariable,
   type PublishesESQLVariable,
   type PublishesESQLVariables,
