+++ src/platform/plugins/shared/controls/public/controls/esql_control/get_esql_control_factory.test.tsx
@@ -11,7 +11,7 @@ import React from 'react';
 import { BehaviorSubject } from 'rxjs';
 import { StateComparators } from '@kbn/presentation-publishing';
 import { fireEvent, render, waitFor } from '@testing-library/react';
-import type { ESQLControlState } from '@kbn/esql/public';
+import type { ESQLControlState } from '@kbn/esql-types';
 import { getMockedControlGroupApi } from '../mocks/control_mocks';
 import type { ControlApiRegistration } from '../types';
 import { getESQLControlFactory } from './get_esql_control_factory';
