+++ src/platform/plugins/shared/controls/public/controls/esql_control/get_esql_control_factory.tsx
@@ -12,10 +12,9 @@ import { i18n } from '@kbn/i18n';
 import { BehaviorSubject } from 'rxjs';
 import { css } from '@emotion/react';
 import { EuiComboBox } from '@elastic/eui';
-import { apiPublishesESQLVariables } from '@kbn/esql-types';
+import { apiPublishesESQLVariables, type ESQLControlState } from '@kbn/esql-types';
 import { useBatchedPublishingSubjects, apiHasParentApi } from '@kbn/presentation-publishing';
 import { tracksOverlays } from '@kbn/presentation-containers';
-import type { ESQLControlState } from '@kbn/esql/public';
 import { ESQL_CONTROL } from '../../../common';
 import type { ESQLControlApi } from './types';
 import { ControlFactory } from '../types';
