+++ src/platform/plugins/shared/controls/tsconfig.json
@@ -39,7 +39,6 @@
     "@kbn/shared-ux-utility",
     "@kbn/std",
     "@kbn/react-hooks",
-    "@kbn/esql",
     "@kbn/esql-types",
   ],
   "exclude": ["target/**/*"]
