+++ src/platform/plugins/shared/esql/public/index.ts
@@ -10,7 +10,6 @@
 import { EsqlPlugin, type EsqlPluginStart } from './plugin';
 
 export { ESQLLangEditor } from './create_editor';
-export { type ESQLControlState, EsqlControlType } from './triggers/esql_controls/types';
 export type { ESQLEditorProps } from '@kbn/esql-editor';
 export type { EsqlPluginStart };
