+++ src/platform/plugins/shared/esql/public/triggers/esql_controls/control_flyout/helpers.ts
@@ -7,9 +7,8 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 import { monaco } from '@kbn/monaco';
-import { ESQLVariableType, ESQLControlVariable } from '@kbn/esql-types';
+import { ESQLVariableType, ESQLControlVariable, VariableNamePrefix } from '@kbn/esql-types';
 import { timeUnits } from '@kbn/esql-validation-autocomplete';
-import { VariableNamePrefix } from '../types';
 
 function inKnownTimeInterval(timeIntervalUnit: string): boolean {
   return timeUnits.some((unit) => unit === timeIntervalUnit.toLowerCase());
