+++ src/platform/plugins/shared/esql/public/triggers/esql_controls/control_flyout/identifier_control_form.tsx
@@ -20,13 +20,16 @@ import {
 } from '@elastic/eui';
 import { monaco } from '@kbn/monaco';
 import type { ISearchGeneric } from '@kbn/search-types';
-import { ESQLVariableType } from '@kbn/esql-types';
+import {
+  ESQLVariableType,
+  EsqlControlType,
+  type ESQLControlState,
+  type ControlWidthOptions,
+} from '@kbn/esql-types';
 import { aggFunctionDefinitions } from '@kbn/esql-validation-autocomplete';
 import { getESQLQueryColumnsRaw } from '@kbn/esql-utils';
-import type { ESQLControlState, ControlWidthOptions } from '../types';
 import { ControlWidth, ControlLabel } from './shared_form_components';
 import { getQueryForFields } from './helpers';
-import { EsqlControlType } from '../types';
 
 interface IdentifierControlFormProps {
   search: ISearchGeneric;
