+++ src/platform/plugins/shared/esql/public/triggers/esql_controls/esql_control_action.ts
@@ -11,10 +11,9 @@ import { i18n } from '@kbn/i18n';
 import type { Action } from '@kbn/ui-actions-plugin/public';
 import type { CoreStart } from '@kbn/core/public';
 import type { ISearchGeneric } from '@kbn/search-types';
-import type { ESQLVariableType } from '@kbn/esql-types';
+import type { ESQLVariableType, ESQLControlState } from '@kbn/esql-types';
 import type { ESQLControlVariable } from '@kbn/esql-types';
 import { monaco } from '@kbn/monaco';
-import type { ESQLControlState } from './types';
 import { isActionCompatible, executeAction } from './esql_control_helpers';
 import { ACTION_CREATE_ESQL_CONTROL } from '../constants';
