+++ src/platform/plugins/shared/esql/public/triggers/esql_controls/esql_control_helpers.tsx
@@ -12,11 +12,10 @@ import { KibanaContextProvider } from '@kbn/kibana-react-plugin/public';
 import { KibanaRenderContextProvider } from '@kbn/react-kibana-context-render';
 import type { CoreStart } from '@kbn/core/public';
 import type { ISearchGeneric } from '@kbn/search-types';
-import type { ESQLVariableType, ESQLControlVariable } from '@kbn/esql-types';
+import type { ESQLVariableType, ESQLControlVariable, ESQLControlState } from '@kbn/esql-types';
 import { toMountPoint } from '@kbn/react-kibana-mount';
 import { monaco } from '@kbn/monaco';
 import { untilPluginStartServicesReady } from '../../kibana_services';
-import type { ESQLControlState } from './types';
 
 interface Context {
   queryString: string;
